+++ CHANGES.md
@@ -1,5 +1,12 @@
 # RIDL Changelog
 
+## v3.1.7.1 - 2022-03-XX
+New features: 
+ - KoBo mayor upgrade [#765](https://github.com/okfn/ckanext-unhcr/pull/765)
+   - Allow selecting formats and filters for KoBo resources before creating a new Dataset from KoBo.
+   - Triggers a resource update If it changes its kobo filters (each resource can change filters and sync with KoBo independently)
+   - Moved KoBo information from packages to the resource level. Now we can handle the relationship with KoBo within a resource.
+
 ## v3.1.7 - 2022-03-14 - STABLE PRODUCTION RELEASE
 Deployment of RIDL v3.1.7 to production environment